chore: add protected route group
