+++ front-end/app/(protected)/brand/[id]/page.tsx
