+++ front-end/app/(protected)/page.tsx
