--- front-end/app/app/layout.tsx
@@ -1,21 +1,0 @@
-import type { Metadata } from "next";
-import "../globals.css";
-import { ThemeProvider } from "@/components/theme-provider";
-
-export const metadata: Metadata = {
-  title: "brandOS",
-  description: "AI-powered brand creation and growth platform",
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className="min-h-screen antialiased">
-        <ThemeProvider>
-          {children}
-        </ThemeProvider>
-      </body>
-    </html>
-  );
-}
-
